+++ src/components/Pages/ComedienRoles/ComedienRoles.tsx
@@ -4,8 +4,8 @@ import "./ComedienRoles.scss";
 
 export const ComedienRoles: React.FC = () => {
     return (
-        <>
+        <div className="pages-container comedien-roles-attribution">
             ComedienRoles
-        </>
+        </div>
     );
 };