+++ src/components/Pages/PlayerName/PlayerName.tsx
@@ -4,8 +4,8 @@ import "./PlayerName.scss";
 
 export const PlayerName: React.FC = () => {
     return (
-        <>
+        <div className="pages-container player-name-attribution">
             PlayerName
-        </>
+        </div>
     );
 };